fix: don't crash when opening participant context menu from the CallParticipantList component
