+++ packages/react-sdk/src/core/components/ParticipantView/index.ts
@@ -1,3 +1,4 @@
+export * from './ParticipantActionsContextMenu';
 export * from './ParticipantView';
 export * from './ParticipantViewContext';
 export * from './DefaultParticipantViewUI';
